refactor(tui): extract generic CrewAIPipelineTUI; rename phase→task
- Move TUI machinery into tools/tui/__init__.py as CrewAIPipelineTUI(App)
  with crew, task_map, record_prefix, verbose, and dry_run injected at
  construction time; callbacks are now instance methods rather than
  module-level closures
- tui.py becomes a 19-line wrapper: CybersquadTUI(CrewAIPipelineTUI) sets
  CSS_PATH, builds the crew, and passes record_prefix="cybersquad"
- Add TODO at CSS_PATH for future theming via constructor path parameter
- Rename SquadMember.phase → .task across squad/__init__.py and all six
  squad member files
- Rename squad_phases() → crew_tasks() in crew.py to match
- --dry-run now applies to TUI mode too: the TUI composes the pipeline
  view and shows "Dry run mode: pipeline not started." instead of
  launching the crew
- Exclude tui.py and tools/tui/* from coverage (Textual apps need a
  display; the 90% floor is maintained at 90.75% without them)
