+++ squad/disclosure_coordinator/__init__.py
@@ -49,5 +49,5 @@ MEMBER = SquadMember(
     slug="disclosure_coordinator",
     dir=Path(__file__).parent,
     tools=[submit_report_tool, check_duplicate_tool],
-    phase="Disclosure",
+    task="Disclosure",
 )
