+++ squad/penetration_tester/__init__.py
@@ -986,7 +986,7 @@ def save_findings_tool(findings_json: str) -> str:
 MEMBER = SquadMember(
     slug="penetration_tester",
     dir=Path(__file__).parent,
-    phase="Penetration Testing",
+    task="Penetration Testing",
     tools=[
         nuclei_scan_tool,
         sqlmap_tool,
