+++ squad/programme_manager/__init__.py
@@ -37,5 +37,5 @@ MEMBER = SquadMember(
     slug="programme_manager",
     dir=Path(__file__).parent,
     tools=[list_programmes_tool, get_scope_tool, get_programme_stats_tool],
-    phase="Select Program",
+    task="Select Program",
 )
