+++ squad/technical_author/__init__.py
@@ -56,5 +56,5 @@ MEMBER = SquadMember(
     slug="technical_author",
     dir=Path(__file__).parent,
     tools=[create_report_tool, calculate_cvss_tool],
-    phase="Reporting",
+    task="Reporting",
 )
