+++ squad/vulnerability_researcher/__init__.py
@@ -74,5 +74,5 @@ MEMBER = SquadMember(
     slug="vulnerability_researcher",
     dir=Path(__file__).parent,
     tools=[triage_tool, lookup_cve_tool, check_duplicate_tool],
-    phase="Vulnerability Research",
+    task="Vulnerability Research",
 )
